+++ java-io-base/src/main/java/nbbrd/io/text/TextResource.java
@@ -7,6 +7,7 @@ import java.io.*;
 import java.nio.charset.Charset;
 import java.nio.charset.CharsetDecoder;
 import java.nio.charset.CharsetEncoder;
+import java.nio.charset.CodingErrorAction;
 import java.util.Optional;
 
 import static nbbrd.io.Resource.newInputStream;
@@ -15,20 +16,34 @@ import static nbbrd.io.Resource.newInputStream;
 public class TextResource {
 
     /**
-     * @deprecated use {@link #newBufferedReader(Class, String, CharsetDecoder)} instead.
+     * @deprecated use {@link #newBufferedReader(Class, String, Charset)} instead.
      */
     @Deprecated
     public @NonNull Optional<BufferedReader> getResourceAsBufferedReader(@NonNull Class<?> anchor, @NonNull String name, @NonNull Charset charset) {
         return Resource.getResourceAsStream(anchor, name)
-                .map(stream -> newBufferedReader(stream, charset.newDecoder()));
+                .map(stream -> newBufferedReader(stream, charset));
+    }
+
+    public @NonNull BufferedReader newBufferedReader(@NonNull Class<?> anchor, @NonNull String name, @NonNull Charset charset) throws IOException {
+        return newBufferedReader(newInputStream(anchor, name), charset);
     }
 
     public @NonNull BufferedReader newBufferedReader(@NonNull Class<?> anchor, @NonNull String name, @NonNull CharsetDecoder decoder) throws IOException {
-        return newBufferedReader(newInputStream(anchor, name), decoder);
+        return newBufferedReader(newInputStream(anchor, name), configureDecoderForInputStreamReader(decoder));
+    }
+
+    public @NonNull BufferedReader newBufferedReader(@NonNull InputStream stream, @NonNull Charset charset) {
+        return new BufferedReader(new InputStreamReader(stream, charset));
     }
 
     public @NonNull BufferedReader newBufferedReader(@NonNull InputStream stream, @NonNull CharsetDecoder decoder) {
-        return new BufferedReader(new InputStreamReader(stream, decoder));
+        return new BufferedReader(new InputStreamReader(stream, configureDecoderForInputStreamReader(decoder)));
+    }
+
+    private @NonNull CharsetDecoder configureDecoderForInputStreamReader(@NonNull CharsetDecoder decoder) {
+        return decoder
+                .onMalformedInput(CodingErrorAction.REPLACE)
+                .onUnmappableCharacter(CodingErrorAction.REPLACE);
     }
 
     public @NonNull BufferedWriter newBufferedWriter(@NonNull OutputStream stream, @NonNull CharsetEncoder encoder) {